+++ template/src/brick.cpp
@@ -96,7 +96,7 @@ Hitbox Brick::get_hitbox() const
     vec2 position = mc.position;
     position.x -= width / 2;
     position.y += width / 2;
-    Square square(position, width);
+    Square square(position, (int)width);
     squares[0] = square;
     Hitbox hitbox({}, squares);
     return hitbox;
